moves ai cli tests to dlthub
