+++ tests/workspace/cli/dlthub/ai/cases/mock_toolkit/.claude-plugin/plugin.json
