+++ tests/workspace/cli/dlthub/ai/cases/mock_toolkit/commands/bootstrap.md
