+++ tests/workspace/cli/dlthub/ai/cases/mock_toolkit/mcp.json
