+++ tests/workspace/cli/dlthub/ai/cases/mock_toolkit/rules/coding.md
