+++ tests/workspace/cli/dlthub/ai/cases/mock_toolkit/skills/find-source/SKILL.md
