+++ tests/workspace/cli/dlthub/ai/cases/mock_workbench/bootstrap/.claude-plugin/plugin.json
