+++ tests/workspace/cli/dlthub/ai/cases/mock_workbench/bootstrap/.claude-plugin/toolkit.json
