+++ tests/workspace/cli/dlthub/ai/cases/mock_workbench/init/.claude-plugin/plugin.json
