+++ tests/workspace/cli/dlthub/ai/cases/mock_workbench/init/AGENTS.md
