+++ tests/workspace/cli/dlthub/ai/cases/mock_workbench/init/rules/base.md
