+++ tests/workspace/cli/dlthub/ai/cases/mock_workbench/rest-api-pipeline/mcp.json
