+++ tests/workspace/cli/dlthub/ai/cases/mock_workbench/rest-api-pipeline/skills/find-source/SKILL.md
