+++ tests/workspace/cli/dlthub/ai/cases/mock_workbench/sql-database/.claude-plugin/toolkit.json
