+++ tests/workspace/cli/dlthub/ai/cases/versioned_workbench/init/.claude-plugin/plugin.json
