+++ tests/workspace/cli/dlthub/ai/cases/versioned_workbench/init/rules/base.md
