+++ tests/workspace/cli/dlthub/ai/cases/versioned_workbench/my-toolkit/.claude-plugin/plugin.json
