+++ tests/workspace/cli/dlthub/ai/cases/versioned_workbench/my-toolkit/rules/coding.md
